+++ Demo/src/main/java/com/braintreepayments/demo/PayPalRequestFactory.java
@@ -2,11 +2,13 @@ package com.braintreepayments.demo;
 
 import android.content.Context;
 
+import com.braintreepayments.api.core.PostalAddress;
 import com.braintreepayments.api.paypal.PayPalCheckoutRequest;
+import com.braintreepayments.api.paypal.PayPalLandingPageType;
 import com.braintreepayments.api.paypal.PayPalPaymentIntent;
+import com.braintreepayments.api.paypal.PayPalPaymentUserAction;
 import com.braintreepayments.api.paypal.PayPalRequest;
 import com.braintreepayments.api.paypal.PayPalVaultRequest;
-import com.braintreepayments.api.core.PostalAddress;
 
 public class PayPalRequestFactory {
 
@@ -17,8 +19,6 @@ public class PayPalRequestFactory {
 
         PayPalVaultRequest request = new PayPalVaultRequest(true);
 
-        boolean useAppLink = Settings.getPayPalLinkType(context).equals(context.getString(R.string.paypal_app_link));
-        request.setAppLinkEnabled(useAppLink);
         if (!buyerEmailAddress.isEmpty()) {
             request.setUserAuthenticationEmail(buyerEmailAddress);
         }
@@ -27,9 +27,9 @@ public class PayPalRequestFactory {
 
         String landingPageType = Settings.getPayPalLandingPageType(context);
         if (context.getString(R.string.paypal_landing_page_type_billing).equals(landingPageType)) {
-            request.setLandingPageType(PayPalRequest.LANDING_PAGE_TYPE_BILLING);
+            request.setLandingPageType(PayPalLandingPageType.LANDING_PAGE_TYPE_BILLING);
         } else if (context.getString(R.string.paypal_landing_page_type_login).equals(landingPageType)) {
-            request.setLandingPageType(PayPalRequest.LANDING_PAGE_TYPE_LOGIN);
+            request.setLandingPageType(PayPalLandingPageType.LANDING_PAGE_TYPE_LOGIN);
         }
 
         if (Settings.isPayPalCreditOffered(context)) {
@@ -58,8 +58,6 @@ public class PayPalRequestFactory {
     ) {
         PayPalCheckoutRequest request = new PayPalCheckoutRequest(amount, true);
 
-        boolean useAppLink = Settings.getPayPalLinkType(context).equals(context.getString(R.string.paypal_app_link));
-        request.setAppLinkEnabled(useAppLink);
         if (!buyerEmailAddress.isEmpty()) {
             request.setUserAuthenticationEmail(buyerEmailAddress);
         }
@@ -68,9 +66,9 @@ public class PayPalRequestFactory {
 
         String landingPageType = Settings.getPayPalLandingPageType(context);
         if (context.getString(R.string.paypal_landing_page_type_billing).equals(landingPageType)) {
-            request.setLandingPageType(PayPalRequest.LANDING_PAGE_TYPE_BILLING);
+            request.setLandingPageType(PayPalLandingPageType.LANDING_PAGE_TYPE_BILLING);
         } else if (context.getString(R.string.paypal_landing_page_type_login).equals(landingPageType)) {
-            request.setLandingPageType(PayPalRequest.LANDING_PAGE_TYPE_LOGIN);
+            request.setLandingPageType(PayPalLandingPageType.LANDING_PAGE_TYPE_LOGIN);
         }
 
         String intentType = Settings.getPayPalIntentType(context);
@@ -83,7 +81,7 @@ public class PayPalRequestFactory {
         }
 
         if (Settings.isPayPalUseractionCommitEnabled(context)) {
-            request.setUserAction(PayPalCheckoutRequest.USER_ACTION_COMMIT);
+            request.setUserAction(PayPalPaymentUserAction.USER_ACTION_COMMIT);
         }
 
         if (Settings.usePayPalAddressOverride(context)) {