+++ PayPal/src/test/java/com/braintreepayments/api/paypal/PayPalCheckoutRequestUnitTest.java
@@ -5,10 +5,6 @@ import android.os.Parcel;
 import com.braintreepayments.api.core.Authorization;
 import com.braintreepayments.api.core.Configuration;
 import com.braintreepayments.api.core.PostalAddress;
-import com.braintreepayments.api.paypal.PayPalCheckoutRequest;
-import com.braintreepayments.api.paypal.PayPalLineItem;
-import com.braintreepayments.api.paypal.PayPalPaymentIntent;
-import com.braintreepayments.api.paypal.PayPalRequest;
 
 import org.json.JSONException;
 import org.junit.Test;
@@ -41,7 +37,7 @@ public class PayPalCheckoutRequestUnitTest {
         assertNull(request.getLandingPageType());
         assertNull(request.getBillingAgreementDescription());
         assertFalse(request.getShouldOfferPayLater());
-        assertFalse(request.hasUserLocationConsent());
+        assertFalse(request.getHasUserLocationConsent());
     }
 
     @Test
@@ -57,10 +53,10 @@ public class PayPalCheckoutRequestUnitTest {
         request.setBillingAgreementDescription("Billing Agreement Description");
         request.setShippingAddressRequired(true);
         request.setShippingAddressOverride(postalAddress);
-        request.setUserAction(PayPalCheckoutRequest.USER_ACTION_COMMIT);
+        request.setUserAction(PayPalPaymentUserAction.USER_ACTION_COMMIT);
         request.setDisplayName("Display Name");
         request.setRiskCorrelationId("123-correlation");
-        request.setLandingPageType(PayPalRequest.LANDING_PAGE_TYPE_LOGIN);
+        request.setLandingPageType(PayPalLandingPageType.LANDING_PAGE_TYPE_LOGIN);
 
         assertEquals("1.00", request.getAmount());
         assertEquals("USD", request.getCurrencyCode());
@@ -70,12 +66,12 @@ public class PayPalCheckoutRequestUnitTest {
         assertTrue(request.isShippingAddressRequired());
         assertEquals(postalAddress, request.getShippingAddressOverride());
         assertEquals(PayPalPaymentIntent.SALE, request.getIntent());
-        assertEquals(PayPalCheckoutRequest.USER_ACTION_COMMIT, request.getUserAction());
+        assertEquals(PayPalPaymentUserAction.USER_ACTION_COMMIT, request.getUserAction());
         assertEquals("Display Name", request.getDisplayName());
         assertEquals("123-correlation", request.getRiskCorrelationId());
-        assertEquals(PayPalRequest.LANDING_PAGE_TYPE_LOGIN, request.getLandingPageType());
+        assertEquals(PayPalLandingPageType.LANDING_PAGE_TYPE_LOGIN, request.getLandingPageType());
         assertTrue(request.getShouldOfferPayLater());
-        assertTrue(request.hasUserLocationConsent());
+        assertTrue(request.getHasUserLocationConsent());
     }
 
     @Test
@@ -92,14 +88,14 @@ public class PayPalCheckoutRequestUnitTest {
         request.setShippingAddressOverride(postalAddress);
 
         request.setIntent(PayPalPaymentIntent.SALE);
-        request.setLandingPageType(PayPalRequest.LANDING_PAGE_TYPE_LOGIN);
-        request.setUserAction(PayPalCheckoutRequest.USER_ACTION_COMMIT);
+        request.setLandingPageType(PayPalLandingPageType.LANDING_PAGE_TYPE_LOGIN);
+        request.setUserAction(PayPalPaymentUserAction.USER_ACTION_COMMIT);
         request.setDisplayName("Display Name");
         request.setRiskCorrelationId("123-correlation");
         request.setMerchantAccountId("merchant_account_id");
 
         ArrayList<PayPalLineItem> lineItems = new ArrayList<>();
-        lineItems.add(new PayPalLineItem(PayPalLineItem.KIND_DEBIT, "An Item", "1", "1"));
+        lineItems.add(new PayPalLineItem(PayPalLineItemKind.DEBIT, "An Item", "1", "1"));
         request.setLineItems(lineItems);
 
         Parcel parcel = Parcel.obtain();
@@ -114,17 +110,16 @@ public class PayPalCheckoutRequestUnitTest {
                 result.getBillingAgreementDescription());
         assertTrue(result.isShippingAddressRequired());
         assertTrue(result.isShippingAddressEditable());
-        assertEquals("Postal Address", result.getShippingAddressOverride()
-                .getRecipientName());
+        assertEquals("Postal Address", result.getShippingAddressOverride().getRecipientName());
         assertEquals(PayPalPaymentIntent.SALE, result.getIntent());
-        assertEquals(PayPalRequest.LANDING_PAGE_TYPE_LOGIN, result.getLandingPageType());
-        assertEquals(PayPalCheckoutRequest.USER_ACTION_COMMIT, result.getUserAction());
+        assertEquals(PayPalLandingPageType.LANDING_PAGE_TYPE_LOGIN, result.getLandingPageType());
+        assertEquals(PayPalPaymentUserAction.USER_ACTION_COMMIT, result.getUserAction());
         assertEquals("Display Name", result.getDisplayName());
         assertEquals("123-correlation", result.getRiskCorrelationId());
         assertEquals("merchant_account_id", result.getMerchantAccountId());
         assertEquals(1, result.getLineItems().size());
         assertEquals("An Item", result.getLineItems().get(0).getName());
-        assertTrue(result.hasUserLocationConsent());
+        assertTrue(result.getHasUserLocationConsent());
     }
 
     @Test