+++ PayPal/src/test/java/com/braintreepayments/api/paypal/PayPalVaultRequestUnitTest.java
@@ -5,9 +5,6 @@ import android.os.Parcel;
 import com.braintreepayments.api.core.Authorization;
 import com.braintreepayments.api.core.Configuration;
 import com.braintreepayments.api.core.PostalAddress;
-import com.braintreepayments.api.paypal.PayPalLineItem;
-import com.braintreepayments.api.paypal.PayPalRequest;
-import com.braintreepayments.api.paypal.PayPalVaultRequest;
 
 import org.json.JSONException;
 import org.junit.Test;
@@ -35,8 +32,7 @@ public class PayPalVaultRequestUnitTest {
         assertNull(request.getDisplayName());
         assertNull(request.getLandingPageType());
         assertFalse(request.getShouldOfferCredit());
-        assertFalse(request.hasUserLocationConsent());
-        assertFalse(request.isAppLinkEnabled());
+        assertFalse(request.getHasUserLocationConsent());
     }
 
     @Test
@@ -49,9 +45,8 @@ public class PayPalVaultRequestUnitTest {
         request.setShippingAddressOverride(postalAddress);
         request.setDisplayName("Display Name");
         request.setRiskCorrelationId("123-correlation");
-        request.setLandingPageType(PayPalRequest.LANDING_PAGE_TYPE_LOGIN);
+        request.setLandingPageType(PayPalLandingPageType.LANDING_PAGE_TYPE_LOGIN);
         request.setShouldOfferCredit(true);
-        request.setAppLinkEnabled(true);
 
         assertEquals("US", request.getLocaleCode());
         assertEquals("Billing Agreement Description", request.getBillingAgreementDescription());
@@ -59,10 +54,9 @@ public class PayPalVaultRequestUnitTest {
         assertEquals(postalAddress, request.getShippingAddressOverride());
         assertEquals("Display Name", request.getDisplayName());
         assertEquals("123-correlation", request.getRiskCorrelationId());
-        assertEquals(PayPalRequest.LANDING_PAGE_TYPE_LOGIN, request.getLandingPageType());
+        assertEquals(PayPalLandingPageType.LANDING_PAGE_TYPE_LOGIN, request.getLandingPageType());
         assertTrue(request.getShouldOfferCredit());
-        assertTrue(request.hasUserLocationConsent());
-        assertTrue(request.isAppLinkEnabled());
+        assertTrue(request.getHasUserLocationConsent());
     }
 
     @Test
@@ -73,19 +67,18 @@ public class PayPalVaultRequestUnitTest {
         request.setShippingAddressRequired(true);
         request.setShippingAddressEditable(true);
         request.setShouldOfferCredit(true);
-        request.setAppLinkEnabled(true);
 
         PostalAddress postalAddress = new PostalAddress();
         postalAddress.setRecipientName("Postal Address");
         request.setShippingAddressOverride(postalAddress);
 
-        request.setLandingPageType(PayPalRequest.LANDING_PAGE_TYPE_LOGIN);
+        request.setLandingPageType(PayPalLandingPageType.LANDING_PAGE_TYPE_LOGIN);
         request.setDisplayName("Display Name");
         request.setRiskCorrelationId("123-correlation");
         request.setMerchantAccountId("merchant_account_id");
 
         ArrayList<PayPalLineItem> lineItems = new ArrayList<>();
-        lineItems.add(new PayPalLineItem(PayPalLineItem.KIND_DEBIT, "An Item", "1", "1"));
+        lineItems.add(new PayPalLineItem(PayPalLineItemKind.DEBIT, "An Item", "1", "1"));
         request.setLineItems(lineItems);
 
         Parcel parcel = Parcel.obtain();
@@ -99,16 +92,14 @@ public class PayPalVaultRequestUnitTest {
         assertTrue(result.getShouldOfferCredit());
         assertTrue(result.isShippingAddressRequired());
         assertTrue(result.isShippingAddressEditable());
-        assertEquals("Postal Address", result.getShippingAddressOverride()
-            .getRecipientName());
-        assertEquals(PayPalRequest.LANDING_PAGE_TYPE_LOGIN, result.getLandingPageType());
+        assertEquals("Postal Address", result.getShippingAddressOverride().getRecipientName());
+        assertEquals(PayPalLandingPageType.LANDING_PAGE_TYPE_LOGIN, result.getLandingPageType());
         assertEquals("Display Name", result.getDisplayName());
         assertEquals("123-correlation", result.getRiskCorrelationId());
         assertEquals("merchant_account_id", result.getMerchantAccountId());
         assertEquals(1, result.getLineItems().size());
         assertEquals("An Item", result.getLineItems().get(0).getName());
-        assertTrue(result.hasUserLocationConsent());
-        assertTrue(result.isAppLinkEnabled());
+        assertTrue(result.getHasUserLocationConsent());
     }
 
     @Test